ickoxii/better-names

+++ mock-data/clear-db.py
@@ -3,10 +3,13 @@ import requests
 import shutil
 import sys
 
-# Production - Brendon's Runner -- USE WITH CAUTION
+# Production - Brendon's Runner
 # API_BASE_URL = "http://35.184.141.85:8080"
 
-# Remote Dev - Icko's Runner
+# Icko's First Runner
+# API_BASE_URL = "http://http://35.224.27.57:8080"
+
+# Remote Dev - Icko's woofrunner
 # API_BASE_URL = "http://35.184.141.85:8080"
 
 # Local Development
@@ -39,33 +42,35 @@ if confirmation == 'yes':
 # <<<< Clear Tables <<<<
 
 # >>>> Clear Mock Uploads >>>>
-confirmation = input("Would you like to clear mock-data/uploads? (yes/no): ").lower()
+if API_BASE_URL == "http://localhost:8080"
+    confirmation = input("Would you like to clear mock-data/uploads? (yes/no): ").lower()
 
-if confirmation == 'yes':
-    print("Removing uploads directory if it exists")
-    uploads_dir = "uploads"
-    if os.path.exists(uploads_dir):
-        try:
-            shutil.rmtree(uploads_dir)
-            print(f"Successfully removed {uploads_dir} directory")
-        except Exception as e:
-            print(f"Error removing {uploads_dir} directory: {e}")
-    else:
-        print(f"{uploads_dir} directory does not exist")
+    if confirmation == 'yes':
+        print("Removing uploads directory if it exists")
+        uploads_dir = "uploads"
+        if os.path.exists(uploads_dir):
+            try:
+                shutil.rmtree(uploads_dir)
+                print(f"Successfully removed {uploads_dir} directory")
+            except Exception as e:
+                print(f"Error removing {uploads_dir} directory: {e}")
+        else:
+            print(f"{uploads_dir} directory does not exist")
 # <<<< Clear Mock Uploads <<<<
 
 # <<<< Clear pet-adoption-api/uploads <<<<
-confirmation = input("Would you like to clear pet-adoption-api/uploads? (yes/no): ").lower()
+if API_BASE_URL == "http://localhost:8080"
+    confirmation = input("Would you like to clear pet-adoption-api/uploads? (yes/no): ").lower()
 
-if confirmation == 'yes':
-    print("Removing uploads directory if it exists")
-    uploads_dir = "../pet-adoption-api/uploads"
-    if os.path.exists(uploads_dir):
-        try:
-            shutil.rmtree(uploads_dir)
-            print(f"Successfully removed {uploads_dir} directory")
-        except Exception as e:
-            print(f"Error removing {uploads_dir} directory: {e}")
-    else:
-        print(f"{uploads_dir} directory does not exist")
+    if confirmation == 'yes':
+        print("Removing uploads directory if it exists")
+        uploads_dir = "../pet-adoption-api/uploads"
+        if os.path.exists(uploads_dir):
+            try:
+                shutil.rmtree(uploads_dir)
+                print(f"Successfully removed {uploads_dir} directory")
+            except Exception as e:
+                print(f"Error removing {uploads_dir} directory: {e}")
+        else:
+            print(f"{uploads_dir} directory does not exist")
 # <<<< Clear pet-adoption-api/uploads <<<<
